made toekn listerner print out each field on a separate line.
Next.. pass in and object which holds the scope variables.. converm that it the same output.

Next.. put that into an interior class wrapped by, say, table manager.

move the output into a  writer class (extend abstract class).

put into this class a writer that takes teh scope and value as arguments..
basically everything that gets printed.

Get it to print out into the same format. Compare files.

Make another class to print the same thing, with, say, a space between the table, the column name, and the value.

Next make a different type of writer that keep tracks of what table we are in.. perhaps the sign for a new table involveds the scope changes? (make a function that does both the scope changes and raises a method that changes the table).

If it is printing decently.. next try printing into separate files (an (ordered?) hashmap with the table name as the key, and part of the target file name).

See if this prints. Perhaps compare a table based on a grep of the previous table. At this point, we haven't written into tables.

ossibly in the first pass of a new table type,

Possibly this ends up similar to programming with stubs..